refactor: Fix type for font measurements

+++ src/Frontend/Resource/Font/WordSizer/WordSizerInterface.php
@@ -13,7 +13,7 @@ namespace PdfGenerator\Frontend\Resource\Font\WordSizer;
 
 interface WordSizerInterface
 {
-    public function getWidth(string $word): float;
+    public function getWidth(string $word): int;
 
-    public function getSpaceWidth(): float;
+    public function getSpaceWidth(): int;
 }
